refactor: move conv neuron plugins
