+++ marble/marblemain.py
@@ -96,192 +96,13 @@ __all__ += [
 ]
 
 
-# -----------------------------
-# Neuron Plugin: Conv1D (pure Python, parameterized by incoming neurons)
-# -----------------------------
-
-class Conv1DNeuronPlugin:
-    """Pure-Python 1D convolution whose parameters are driven by connected neurons.
-
-    Requires at least 5 incoming synapses to the target neuron; parameters are read
-    from the source neurons' tensors in a deterministic order (sorted by source
-    position when available, else by source id):
-      1) kernel: list[float] (1D filter coefficients)
-      2) stride: int >= 1
-      3) padding: int >= 0 (zero-pad both sides)
-      4) dilation: int >= 1
-      5) bias: float (added to each output element)
-
-    The input signal is taken from the provided input_value (if not None) or the
-    neuron's own tensor, coerced to a flat list[float]. The output is a list[float].
-    """
-
-    def on_init(self, neuron: "Neuron") -> None:
-        # Strict wiring validation at creation time: exactly 5 PARAM inputs and exactly 1 outgoing
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 5 or len(out) != 1:
-            raise ValueError(
-                f"Conv1D neuron requires exactly 5 incoming PARAM synapses and exactly 1 outgoing synapse; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "conv1d_init", {"incoming": len(inc), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def _key_src(self, syn: "Synapse"):
-        src = syn.source
-        pos = getattr(src, "position", None)
-        if isinstance(pos, tuple):
-            return (0, tuple(pos))
-        return (1, id(src))
-
-    def _to_list1d(self, x) -> list:
-        try:
-            if hasattr(x, "detach") and hasattr(x, "tolist"):
-                lst = x.detach().to("cpu").view(-1).tolist()
-            elif isinstance(x, (list, tuple)):
-                lst = list(x)
-            else:
-                lst = [x]
-        except Exception:
-            lst = []
-        out = []
-        for v in lst:
-            try:
-                out.append(float(v))
-            except Exception:
-                out.append(0.0)
-        return out
-
-    def _first_scalar(self, x, *, default: float = 0.0, min_val: Optional[float] = None) -> float:
-        vals = self._to_list1d(x)
-        v = float(vals[0]) if vals else float(default)
-        if min_val is not None and v < min_val:
-            v = float(min_val)
-        return v
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        # Select parameter synapses explicitly when labeled; fallback to first 5
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) >= 5:
-            param_incs.sort(key=self._key_src)
-            sel = param_incs[:5]
-        else:
-            if len(incoming) < 5:
-                raise ValueError("Conv1D plugin requires 5 incoming synapses (kernel, stride, padding, dilation, bias)")
-            incoming.sort(key=self._key_src)
-            sel = incoming[:5]
-        k_src = sel[0].source
-        s_src = sel[1].source
-        p_src = sel[2].source
-        d_src = sel[3].source
-        b_src = sel[4].source
-
-        kernel = self._to_list1d(getattr(k_src, "tensor", [])) or [1.0]
-        stride = int(self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        padding = int(max(0.0, self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)))
-        dilation = int(self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        bias = float(self._first_scalar(getattr(b_src, "tensor", 0.0), default=0.0))
-
-        # Prefer per-neuron learnable parameters if provided by SelfAttention
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        learn_kernel = lstore.get("kernel")
-        learn_bias = lstore.get("conv_bias")
-
-        # Build input from incoming DATA synapses if present; else use input_value/tensor
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            x_list_all: List[float] = []
-            for s in data_incs:
-                x_list_all += self._to_list1d(getattr(s.source, "tensor", []))
-            x1 = x_list_all
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x1 = self._to_list1d(x)
-
-        # If torch is available, perform computation on the neuron's device (CUDA preferred)
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        # Use high-performance torch path only on CUDA; on CPU prefer pure-Python to avoid
-        # backend initialization warnings on some hardware (e.g., NNPACK unsupported).
-        if torch is not None and str(device) == "cuda":
-            try:
-                xt = torch.tensor(x1, dtype=torch.float32, device=device).view(1, 1, -1)
-                wt = (learn_kernel.to(device) if hasattr(learn_kernel, "to") else torch.tensor(kernel, dtype=torch.float32, device=device)).view(1, 1, -1)
-                bt_val = float(bias)
-                bt = (learn_bias.to(device).view(-1) if hasattr(learn_bias, "to") else torch.tensor([bt_val], dtype=torch.float32, device=device))
-                # Use conv1d with provided stride/padding/dilation; this runs on CUDA when available
-                y = torch.nn.functional.conv1d(xt, wt, bias=bt, stride=stride, padding=padding, dilation=dilation)
-                y = y.view(-1)
-                try:
-                    report("neuron", "conv1d", {"in": int(xt.numel()), "out": int(y.numel()), "k": int(wt.numel()), "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                # Fallback to pure Python path below
-                pass
-
-        # Zero-pad and compute in pure Python when torch is unavailable or failed
-        if padding > 0:
-            x1 = ([0.0] * padding) + x1 + ([0.0] * padding)
-        n = len(x1)
-        klen = len(kernel)
-        span = (klen - 1) * dilation + 1
-        out_len = 0
-        if n >= span:
-            out_len = 1 + (n - span) // stride
-        y_list: list = []
-        # If a learnable kernel is present but torch path failed, convert to list
-        if hasattr(learn_kernel, "detach"):
-            try:
-                kernel = list(learn_kernel.detach().to("cpu").view(-1).tolist())
-                klen = len(kernel)
-            except Exception:
-                pass
-        if hasattr(learn_bias, "detach"):
-            try:
-                bias = float(learn_bias.detach().to("cpu").view(-1)[0].item())
-            except Exception:
-                try:
-                    bias = float(learn_bias)
-                except Exception:
-                    pass
-        for t in range(out_len):
-            base = t * stride
-            acc = 0.0
-            for i in range(klen):
-                xi = base + i * dilation
-                acc += kernel[i] * x1[xi]
-            y_list.append(acc + bias)
-        try:
-            report("neuron", "conv1d", {"in": len(x1), "out": len(y_list), "k": klen, "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)  # type: ignore[attr-defined]
-        except Exception:
-            return y_list
-
-
 # Register conv1d plugin via dedicated module
 try:
-    from .plugins.conv1d import Conv1DNeuronPlugin as _Conv1DPlugin
-    register_neuron_type("conv1d", _Conv1DPlugin())
+    from .plugins.conv1d import Conv1DNeuronPlugin
+    register_neuron_type("conv1d", Conv1DNeuronPlugin())
 except Exception:
     pass
 
-__all__ += ["Conv1DNeuronPlugin"]
 
 
 # -----------------------------
@@ -462,343 +283,14 @@ class _ConvNDCommon:
         return v
 
 
-class Conv2DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 5 or len(out) != 1:
-            raise ValueError(
-                f"Conv2D neuron requires exactly 5 incoming PARAM synapses and exactly 1 outgoing synapse; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "conv2d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 5:
-            raise ValueError("Conv2D plugin requires 5 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src, d_src, b_src = [s.source for s in param_incs[:5]]
-
-        kernel_1d = self._to_list1d(getattr(k_src, "tensor", [])) or [1.0]
-        stride = int(self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        padding = int(max(0.0, self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)))
-        dilation = int(self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        bias = float(self._first_scalar(getattr(b_src, "tensor", 0.0), default=0.0))
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        learn_kernel = lstore.get("kernel")
-        learn_bias = lstore.get("conv_bias")
-
-        # Build input from incoming DATA synapses if present; fallback to input_value/tensor
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            rows: List[List[float]] = []
-            for s in data_incs:
-                rows.append(self._to_list1d(getattr(s.source, "tensor", [])))
-            # Use common width (min length) across rows
-            if rows:
-                width = min((len(r) for r in rows if r), default=0)
-            else:
-                width = 0
-            if width <= 0:
-                x_vals = []
-                H = 0
-                W = 0
-            else:
-                rows = [r[:width] for r in rows]
-                H = len(rows)
-                W = width
-                x_vals = [v for r in rows for v in r]
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x_vals = self._to_list1d(x)
-            N = max(1, len(x_vals))
-            rh = int(math.isqrt(N))
-            if rh * rh == N:
-                H = W = rh
-            else:
-                H, W = N, 1
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0 and str(device) == "cuda":
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, H, W)
-                # Infer kernel shape: prefer square if perfect square else (len,1)
-                L = max(1, len(kernel_1d))
-                r = int(math.isqrt(L))
-                if r * r == L:
-                    kh = kw = r
-                else:
-                    kh, kw = L, 1
-                base_w = (learn_kernel.to(device) if hasattr(learn_kernel, "to") else torch.tensor(kernel_1d, dtype=torch.float32, device=device))
-                wt = base_w.view(1, 1, kh, kw)
-                bt = (learn_bias.to(device).view(-1) if hasattr(learn_bias, "to") else torch.tensor([bias], dtype=torch.float32, device=device))
-                y = torch.nn.functional.conv2d(
-                    xt, wt, bias=bt, stride=(stride, stride), padding=(padding, padding), dilation=(dilation, dilation)
-                )
-                y = y.view(-1)
-                try:
-                    report("neuron", "conv2d", {"inHW": [H, W], "out": int(y.numel()), "kHW": [kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure Python fallback
-        if padding > 0 and H > 0 and W > 0:
-            padded = []
-            zero_row = [0.0] * (W + 2 * padding)
-            for _ in range(padding):
-                padded.append(list(zero_row))
-            for r_ in range(H):
-                row = [0.0] * padding + x_vals[r_ * W:(r_ + 1) * W] + [0.0] * padding
-                padded.append(row)
-            for _ in range(padding):
-                padded.append(list(zero_row))
-            H2, W2 = len(padded), len(padded[0])
-        else:
-            padded = [x_vals[r_ * W:(r_ + 1) * W] for r_ in range(H)] if H > 0 and W > 0 else []
-            H2, W2 = (len(padded), len(padded[0])) if padded else (0, 0)
-        # Infer kernel shape
-        # If learnable present but torch path failed, convert to list
-        if hasattr(learn_kernel, "detach"):
-            try:
-                kernel_1d = list(learn_kernel.detach().to("cpu").view(-1).tolist())
-            except Exception:
-                pass
-        if hasattr(learn_bias, "detach"):
-            try:
-                bias = float(learn_bias.detach().to("cpu").view(-1)[0].item())
-            except Exception:
-                try:
-                    bias = float(learn_bias)
-                except Exception:
-                    pass
-        L = max(1, len(kernel_1d))
-        r = int(math.isqrt(L))
-        if r * r == L:
-            kh = kw = r
-        else:
-            kh, kw = L, 1
-        out_h = 0 if H2 < kh else 1 + (H2 - ((kh - 1) * dilation + 1)) // stride
-        out_w = 0 if W2 < kw else 1 + (W2 - ((kw - 1) * dilation + 1)) // stride
-        y_list: List[float] = []
-        for oy in range(out_h):
-            base_y = oy * stride
-            for ox in range(out_w):
-                base_x = ox * stride
-                acc = 0.0
-                for ky in range(kh):
-                    for kx in range(kw):
-                        iy = base_y + ky * dilation
-                        ix = base_x + kx * dilation
-                        acc += kernel_1d[ky * kw + kx] * padded[iy][ix]
-                y_list.append(acc + bias)
-        try:
-            report("neuron", "conv2d", {"inHW": [H, W], "out": len(y_list), "kHW": [kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)
-        except Exception:
-            return y_list
-
-
-class Conv3DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 5 or len(out) != 1:
-            raise ValueError(
-                f"Conv3D neuron requires exactly 5 incoming PARAM synapses and exactly 1 outgoing synapse; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "conv3d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 5:
-            raise ValueError("Conv3D plugin requires 5 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src, d_src, b_src = [s.source for s in param_incs[:5]]
-
-        kernel_1d = self._to_list1d(getattr(k_src, "tensor", [])) or [1.0]
-        stride = int(self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        padding = int(max(0.0, self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)))
-        dilation = int(self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        bias = float(self._first_scalar(getattr(b_src, "tensor", 0.0), default=0.0))
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        learn_kernel = lstore.get("kernel")
-        learn_bias = lstore.get("conv_bias")
-
-        # Build input from incoming DATA synapses if present; fallback to input_value/tensor
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            # Treat each data source as one 2D slice; infer H,W per slice as square if possible else (len,1)
-            slices: List[List[float]] = []
-            dims: List[Tuple[int, int]] = []
-            for s in data_incs:
-                vals = self._to_list1d(getattr(s.source, "tensor", []))
-                N = max(1, len(vals))
-                r = int(math.isqrt(N))
-                if r * r == N:
-                    H = W = r
-                else:
-                    H, W = N, 1
-                dims.append((H, W))
-                slices.append(vals[: H * W])
-            # Harmonize to the smallest H,W across slices
-            if dims:
-                Hmin = min(h for h, _ in dims)
-                Wmin = min(w for _, w in dims)
-            else:
-                Hmin = Wmin = 0
-            x_vals = []
-            D = len(slices)
-            for sl, (h, w) in zip(slices, dims):
-                # reshape sl to h x w then crop to Hmin x Wmin and flatten
-                for rr in range(Hmin):
-                    row = sl[rr * w:(rr + 1) * w]
-                    x_vals.extend(row[:Wmin])
-            H, W = Hmin, Wmin
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            vals = self._to_list1d(x)
-            N = max(1, len(vals))
-            r3 = round(N ** (1.0 / 3.0))
-            if r3 > 0 and (r3 * r3 * r3) == N:
-                D = H = W = int(r3)
-                x_vals = vals[: D * H * W]
-            else:
-                D, H, W = N, 1, 1
-                x_vals = vals
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0 and str(device) == "cuda":
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, D, H, W)
-                # Infer kernel shape: prefer cube if perfect cube; else (L,1,1)
-                L = max(1, len(kernel_1d))
-                r = round(L ** (1.0 / 3.0))
-                if r > 0 and (r * r * r) == L:
-                    kd = kh = kw = int(r)
-                else:
-                    kd, kh, kw = L, 1, 1
-                base_w = (learn_kernel.to(device) if hasattr(learn_kernel, "to") else torch.tensor(kernel_1d, dtype=torch.float32, device=device))
-                wt = base_w.view(1, 1, kd, kh, kw)
-                bt = (learn_bias.to(device).view(-1) if hasattr(learn_bias, "to") else torch.tensor([bias], dtype=torch.float32, device=device))
-                y = torch.nn.functional.conv3d(
-                    xt, wt, bias=bt, stride=(stride, stride, stride), padding=(padding, padding, padding), dilation=(dilation, dilation, dilation)
-                )
-                y = y.view(-1)
-                try:
-                    report("neuron", "conv3d", {"inDHW": [D, H, W], "out": int(y.numel()), "kDHW": [kd, kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure Python fallback
-        if padding > 0 and H > 0 and W > 0:
-            D2, H2, W2 = D + 2 * padding, H + 2 * padding, W + 2 * padding
-            padded = [0.0] * (D2 * H2 * W2)
-            for dz in range(D):
-                for yy in range(H):
-                    for xx in range(W):
-                        pd = dz + padding
-                        py = yy + padding
-                        px = xx + padding
-                        padded[pd * (H2 * W2) + py * W2 + px] = x_vals[dz * (H * W) + yy * W + xx]
-        else:
-            padded = list(x_vals)
-            D2, H2, W2 = D, H, W
-        out_d = 0 if D2 < 1 else 1 + (D2 - ((1 - 1) * dilation + 1)) // stride  # kernel depth inferred below
-        # Infer kernel shape
-        if hasattr(learn_kernel, "detach"):
-            try:
-                kernel_1d = list(learn_kernel.detach().to("cpu").view(-1).tolist())
-            except Exception:
-                pass
-        if hasattr(learn_bias, "detach"):
-            try:
-                bias = float(learn_bias.detach().to("cpu").view(-1)[0].item())
-            except Exception:
-                try:
-                    bias = float(learn_bias)
-                except Exception:
-                    pass
-        L = max(1, len(kernel_1d))
-        r = round(L ** (1.0 / 3.0))
-        if r > 0 and (r * r * r) == L:
-            kd = kh = kw = int(r)
-        else:
-            kd, kh, kw = L, 1, 1
-        def idx3(a, d, h, w, H_, W_):
-            return a[d * (H_ * W_) + h * W_ + w]
-        out_d = 0 if D2 < kd else 1 + (D2 - ((kd - 1) * dilation + 1)) // stride
-        out_h = 0 if H2 < kh else 1 + (H2 - ((kh - 1) * dilation + 1)) // stride
-        out_w = 0 if W2 < kw else 1 + (W2 - ((kw - 1) * dilation + 1)) // stride
-        y_list: List[float] = []
-        for od in range(out_d):
-            base_d = od * stride
-            for oy in range(out_h):
-                base_y = oy * stride
-                for ox in range(out_w):
-                    base_x = ox * stride
-                    acc = 0.0
-                    for kz in range(kd):
-                        for ky in range(kh):
-                            for kx in range(kw):
-                                iz = base_d + kz * dilation
-                                iy = base_y + ky * dilation
-                                ix = base_x + kx * dilation
-                                acc += kernel_1d[(kz * kh + ky) * kw + kx] * idx3(padded, iz, iy, ix, H2, W2)
-                    y_list.append(acc + bias)
-        try:
-            report("neuron", "conv3d", {"inDHW": [D, H, W], "out": len(y_list), "kDHW": [kd, kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)
-        except Exception:
-            return y_list
-
-
 try:
-    from .plugins.conv2d import Conv2DNeuronPlugin as _Conv2DPlugin
-    from .plugins.conv3d import Conv3DNeuronPlugin as _Conv3DPlugin
-    register_neuron_type("conv2d", _Conv2DPlugin())
-    register_neuron_type("conv3d", _Conv3DPlugin())
+    from .plugins.conv2d import Conv2DNeuronPlugin
+    from .plugins.conv3d import Conv3DNeuronPlugin
+    register_neuron_type("conv2d", Conv2DNeuronPlugin())
+    register_neuron_type("conv3d", Conv3DNeuronPlugin())
 except Exception:
     pass
 
-__all__ += ["Conv2DNeuronPlugin", "Conv3DNeuronPlugin"]
 
 
 # -----------------------------
